Add automatic forwarding the applet server to public using forward-bridge-server
so the user doesn't have to have open firewall for specific port and IP address of local network.
The development machine and connected devices doesn't have to be in the same local network to make it work.
The communication is over SSL then so there is no problem with web browser security, like in chrome emulator was before