Fix the default organization uid setting
The issue was, that when the loadConfig function within the updateConfig function
has been called without the options (object containing profile),
then the particular config was duplicated in the .sosrc.file -
so instead of appending only:

defaultOrganizationUid=defaultOrganizationUid5

to the selected profile (i.e. profile "a"), it was appended:

defaultOrganizationUid=defaultOrganizationUid5

[profile a.profile a]
identification=XXX
apiSecurityToken=XXX
apiUrl=https://api.a.dev.signageos.io

By calling the loadConfig function with the options parameter,
only the desired config is loaded and affected and the updateConfig works properly.