Do the initial build of applet when starting hot-reload and wait for built before watching
When an applet is not initially built it's started serving an applet to device too soon and it practically always fails for the first time.
It's very confusing and also, some old devices and platforms cannot simply handle this unexpected state and can stuck forever. Like Tizen 2.4.
Now, it's waiting for the first build and then continue resolving build. So it's connected to device when applet .package.zip is ready.
It has to disable initial build using chokidar watcher, because it would compile twice.